Remove old files, rename V2
