+++ strategies/SimplePatienceStrategy.py
@@ -1,7 +1,7 @@
-from .LinearAdaptivePatienceStrategy import LinearAdaptivePatienceStrategy
+from .IterativeStrategy import IterativeStrategy
 
 
-class SimplePatienceStrategy(LinearAdaptivePatienceStrategy):
+class SimplePatienceStrategy(IterativeStrategy):
     """
     Patience Equation: p(x) = p
     """
@@ -9,10 +9,9 @@ class SimplePatienceStrategy(LinearAdaptivePatienceStrategy):
     _name = "simple_patience"
     _short_name = "SP"
 
-    def __init__(self, patience: int = 0, **kwargs):
-        import math
-
-        super().__init__(b=patience, min_patience=0, max_patience=999999999, **kwargs)
+    def __init__(self, b: int = 0, **kwargs):
+        super().__init__(**kwargs)
+        self.b = b
 
     @property
     def patience(self):
@@ -20,29 +19,14 @@ class SimplePatienceStrategy(LinearAdaptivePatienceStrategy):
         Redefined here because cannot change property
         setter without also redefining property in same class.
         """
-        return super().patience
-
-    @patience.setter
-    def patience(self, new_patience: int):
-        """
-        Note that patience is alias for b in
-        the polynomial patience equation.
-
-        Parameters:
-        ----------
-        new_patience: int
-            The new value for patience.
-        """
-        self.b = new_patience
+        return lambda x: self.b
 
     @classmethod
     def kwargs(cls) -> dict[str, str]:
         kwargs = super().kwargs()
-        kwargs.pop("a", None)
-        kwargs.pop("b", None)
         kwargs.update(
             {
-                "patience": "p",
+                "b": "b",
             }
         )
         return kwargs
